test(cli): split command package tests
