+++ .sampo/changesets/820-split-cli-package-tests.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split create and add package-level CLI tests into command-family files with shared helpers.
